requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
requests
apscheduler
python-telegram-bot
alembic
